Move dummy statellite code into test.
