+++ test/dummy_satellite/dummy_satellite.cpp
