+++ test/dummy_satellite/stream_dummy_satellite.cpp
