+++ test/satellite_connection_test.cpp
@@ -1,4 +1,5 @@
 #include "../src/satellite_connection.h"
+#include "dummy_satellite/dummy_satellite.h"
 
 #define CATCH_CONFIG_MAIN
 #include <lightstep/catch2/catch.hpp>
@@ -6,11 +7,20 @@
 using namespace lightstep;
 
 TEST_CASE("SatelliteConnection") {
-  /* SECTION("Moving from a socket sets an invalid file descriptor") { */
-  /*   Socket s1; */
-  /*   auto file_descriptor = s1.file_descriptor(); */
-  /*   Socket s2{std::move(s1)}; */
-  /*   CHECK(s2.file_descriptor() == file_descriptor); */
-  /*   CHECK(s1.file_descriptor() == -1); */
-  /* } */
+  configuration_proto::TracerConfiguration configuration;
+  configuration.set_collector_host("localhost");
+  configuration.set_collector_port(9001);
+  auto satellite = MakeDummySatellite(configuration);
+  Logger logger;
+
+  SECTION("We can successfully establish a satellite connection.") {
+    SatelliteConnection connection{logger, "localhost", 9001};
+    CHECK(connection.is_connected());
+  }
+
+  SECTION(
+      "If SatelliteConnection fails to establish a connection, is_connected "
+      "returns false") {
+    SatelliteConnection connection{logger, "localhost", 9002};
+  }
 }
